+++ src/Exception/MissingDependencyException.php
@@ -1,0 +1,7 @@
+<?php
+
+declare(strict_types=1);
+
+namespace DK\OpenXml\Exception;
+
+final class MissingDependencyException extends OpenXmlException {}