+++ src/OfficeFileFormat.php
@@ -1,0 +1,13 @@
+<?php
+
+declare(strict_types=1);
+
+namespace DK\OpenXml;
+
+enum OfficeFileFormat
+{
+    case OpcPackage;
+    case EncryptedOpcPackage;
+    case CompoundFile;
+    case Unknown;
+}